Remove top level containers folder and move under Artist
