+++ src/Styleguide/Pages/Artist/Containers/AuctionResults/AuctionResults.tsx
