+++ src/Styleguide/Pages/Artist/Containers/AuctionResults/index.tsx
